+++ cmds/startTimeEntry.mjs
@@ -1,4 +1,4 @@
-import { defaultWorkspaceId,getProjectByName, createTimeEntry, getProjectById, defaultProjectId } from '../utils.js'
+import { defaultWorkspaceId, getProjectByName, createTimeEntry, getProjectById, defaultProjectId, parseTime } from '../utils.js'
 import dayjs from 'dayjs'
 
 export const command = 'start'
@@ -35,7 +35,6 @@ export const handler = async function (argv) {
 
   if (argv.startTime) {
     let startTime;
-    console.log(argv)
     if (dayjs(argv.startTime).isValid()) {
       startTime = argv.startTime
     } else {
@@ -49,28 +48,6 @@ export const handler = async function (argv) {
 
   params.projectId = project?.id || defaultProjectId || null
   // TODO check for invalid projectId or catch the error when creating fails
-  console.log(params)
   const timeEntry = await createTimeEntry(params)
-  console.log(timeEntry)
   console.info(`Started ${timeEntry?.description} ${project?.name ? `for project ${project.name}` : 'without a project'}`)
-}
-
-/**
- * Parses a timelike string into a dayjs object of the current date and that time
- * @param {string} timeString timelike string e.g. 4:50PM '12:00 AM' etc.
- * @returns {object} dayjs object
- */
-function parseTime (timeString) {
-  let h, m
-  // Assumes time in format 4:50 PM
-  const time = timeString.split(':', 2)
-  h = time[0]
-  m = time[1].match(/[0-9]+/)[0]
-  if (timeString.match(/PM/i) && h <= 12) {
-    // + in front of string converts to a number, cool!
-    h = +h + 12
-  } else if (h == 12) {
-    h = 0
-  }
-  return dayjs().hour(h).minute(m).second(0).millisecond(0)
 }
