+++ utils.js
@@ -144,3 +144,23 @@ export const displayTimeEntry = async function (timeEntry) {
     console.info(`Workspace: ${workspace.name} (#${timeEntry.wid})`)
   }
 }
+
+/**
+ * Parses a timelike string into a dayjs object of the current date and that time
+ * @param {string} timeString timelike string e.g. 4:50PM '12:00 AM' etc.
+ * @returns {object} dayjs object
+ */
+export function parseTime (timeString) {
+  let h, m
+  // Assumes time in format 4:50 PM
+  const time = timeString.split(':', 2)
+  h = time[0]
+  m = time[1].match(/[0-9]+/)[0]
+  if (timeString.match(/PM/i) && h <= 12) {
+    // + in front of string converts to a number, cool!
+    h = +h + 12
+  } else if (h == 12) {
+    h = 0
+  }
+  return dayjs().hour(h).minute(m).second(0).millisecond(0)
+}
